+++ OneBlink.SDK/models/FormElementLookupStaticDataPrefill.cs
@@ -1,0 +1,22 @@
+namespace OneBlink.SDK.Model
+{
+    public class FormElementLookupStaticDataPreFill
+    {
+        public string formElementName
+        {
+            get; set;
+        }
+        public string type
+        {
+            get; set;
+        }
+        public string text
+        {
+            get; set;
+        }
+        public long number
+        {
+            get; set;
+        }
+    }
+}