+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.1.12 (2023-07-03)
+
 ### Added
 
 -   `lookupButton` to `FormElement`